+++ Source/ReDX.cpp
@@ -8,12 +8,12 @@ using namespace DirectX;
 
 // Key press status: 1 if pressed, 0 otherwise.
 struct KeyPressStatus {
-    uint w : 1;
-    uint s : 1;
-    uint a : 1;
-    uint d : 1;
-    uint q : 1;
-    uint e : 1;
+    uint32_t w : 1;
+    uint32_t s : 1;
+    uint32_t a : 1;
+    uint32_t d : 1;
+    uint32_t q : 1;
+    uint32_t e : 1;
 };
 
 int __cdecl main(const int argc, const char* argv[]) {
@@ -36,7 +36,9 @@ int __cdecl main(const int argc, const char* argv[]) {
     // Provide the scene description.
     Scene scene{"..\\..\\Assets\\Sponza\\", "sponza.obj", engine};
     // Set up the camera.
-    PerspectiveCamera pCam{Window::width(), Window::height(), VERTICAL_FOV,
+    PerspectiveCamera pCam{static_cast<float>(Window::width()),
+                           static_cast<float>(Window::height()),
+                           VERTICAL_FOV,
                            /* pos */ {300.f, 200.f, -35.f},
                            /* dir */ {-1.f, 0.f, 0.f},
                            /* up  */ {0.f, 1.f, 0.f}};
@@ -44,7 +46,7 @@ int __cdecl main(const int argc, const char* argv[]) {
     KeyPressStatus keyPressStatus{};
     // Initialize the timings.
     float  timeDelta = 0.f;
-    uint64 cpuTime0, gpuTime0;
+    uint64_t cpuTime0, gpuTime0;
     std::tie(cpuTime0, gpuTime0) = engine.getTime();
     // Main loop.
     while (true) {
@@ -59,7 +61,7 @@ int __cdecl main(const int argc, const char* argv[]) {
                 case WM_KEYUP:
                 case WM_KEYDOWN:
                     {
-                        const uint status = msg.message ^ 0x1;
+                        const size_t status = msg.message ^ 0x1;
                         switch (msg.wParam) {
                             case 0x57: keyPressStatus.w = status; break;
                             case 0x53: keyPressStatus.s = status; break;
@@ -107,10 +109,10 @@ int __cdecl main(const int argc, const char* argv[]) {
         engine.renderFrame();
         // Update the timings.
         {
-            uint64 cpuTime1, gpuTime1;
+            uint64_t cpuTime1, gpuTime1;
             std::tie(cpuTime1, gpuTime1) = engine.getTime();
-            const uint64 cpuFrameTime    = cpuTime1 - cpuTime0;
-            const uint64 gpuFrameTime    = gpuTime1 - gpuTime0;
+            const uint64_t cpuFrameTime  = cpuTime1 - cpuTime0;
+            const uint64_t gpuFrameTime  = gpuTime1 - gpuTime0;
             // Convert the frame times from microseconds to milliseconds.
             Window::displayInfo(cpuFrameTime * 1e-3f, gpuFrameTime * 1e-3f);
             // Convert the frame time from microseconds to seconds.
